+++ shardingsphere-observability/shardingsphere-tracing/shardingsphere-tracing-opentracing/src/main/java/org/apache/shardingsphere/tracing/opentracing/constant/ShardingTags.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.opentracing.constant;
+package org.apache.shardingsphere.tracing.opentracing.constant;
 
 import io.opentracing.tag.StringTag;
 import lombok.AccessLevel;
