+++ shardingsphere-observability/shardingsphere-tracing/shardingsphere-tracing-opentracing/src/main/resources/META-INF/services/org.apache.shardingsphere.control.panel.spi.ControlPanelFacade
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.opentracing.OpenTracingTracer
+org.apache.shardingsphere.tracing.opentracing.OpenTracingTracer
