+++ shardingsphere-observability/shardingsphere-tracing/shardingsphere-tracing-opentracing/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.executor.sql.hook.SQLExecutionHook
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.opentracing.hook.OpenTracingSQLExecutionHook
+org.apache.shardingsphere.tracing.opentracing.hook.OpenTracingSQLExecutionHook
