+++ shardingsphere-observability/shardingsphere-tracing/shardingsphere-tracing-opentracing/src/main/resources/META-INF/services/org.apache.shardingsphere.infra.hook.RootInvokeHook
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.opentracing.hook.OpenTracingRootInvokeHook
+org.apache.shardingsphere.tracing.opentracing.hook.OpenTracingRootInvokeHook
