+++ shardingsphere-observability/shardingsphere-tracing/shardingsphere-tracing-opentracing/src/test/java/org/apache/shardingsphere/tracing/opentracing/hook/OpenTracingRootInvokeHookTest.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.opentracing.hook;
+package org.apache.shardingsphere.tracing.opentracing.hook;
 
 import org.apache.shardingsphere.infra.executor.kernel.ExecutorDataMap;
 import org.apache.shardingsphere.infra.hook.RootInvokeHook;
